refactoring in different files, fixed some bugs, added bot default commands

+++ bot/states/FormToJoinAsBuyer.py
@@ -4,7 +4,7 @@ from aiogram.dispatcher.filters.state import State, StatesGroup
 
 class FormToJoinAsBuyer(StatesGroup):
     amount_payed = State()
-    amount_max: int
+    amount_max: float
     message: types.Message
     call: types.CallbackQuery
     purchase_id: str
